requirements.txt
BTrees==4.0.8
Babel==0.9.6
BeautifulSoup==3.2.1
Chameleon==2.11
Cheetah==2.4.4
CherryPy==3.2.4
Cython==0.20.1
DerApproximator==0.52
Django==1.5.1
EasyProcess==0.1.6
Fabric==1.6.1
Fiona==0.15
Flask==0.9
Flask-Babel==0.8
Flask-HTTPAuth==2.2.0
Flask-Login==0.1.3
Flask-Mail==0.8.2
Flask-OpenID==1.1.1
Flask-SQLAlchemy==0.16
Flask-Script==0.5.3
Flask-WTF==0.8.3
Flask-WhooshAlchemy==0.55a
FuncDesigner==0.5303
GDAL==1.9.1
Genshi==0.7
GeoIP==1.2.9
GitPython==0.3.2.RC1
IMAPClient==0.10.2
Jinja2==2.7
Mako==0.8.1
Markdown==2.3.1
MarkupSafe==0.18
Mezzanine==1.4.3
MySQL-python==1.2.4
Orange==2.7.2
Orange-Text==1.2a1
PIL==1.1.7
Paste==1.7.5.1
PasteDeploy==1.5.0
Pattern==2.6
Pillow==2.4.0
Pinax==0.9a2
PyAMF==0.6.1
PyPDF2==1.20
PySAL==1.5.0
PyStemmer==1.3.0
PyVirtualDisplay==0.1.2
PyYAML==3.11
Pygments==1.6
Pykka==1.1.0
Pyste==0.9.10
QuantLib-Python==1.2
RestrictedPython==3.6.0
SPARQLWrapper==1.5.2
SQLAlchemy==0.8.3
Scrapy==0.16.5
Shapely==1.2.17
South==0.8.1
SpaceFuncs==0.52
Sphinx==1.2b1
Tempita==0.5.1
TornadIO2==0.0.3
Trac==1.0.1
Twisted==13.0.0
Unidecode==0.04.13
VTK==5.8.0
WTForms==1.0.4
WeasyPrint==0.21
WebOb==1.2.3
Werkzeug==0.8.3
Whoosh==2.4.1
ZConfig==3.0.3
ZEO==4.0.0b1
ZODB==4.0.0b3
ZODB3==3.11.0a3
alembic==0.6.2
ansible==1.5.4
argparse==1.2.1
async==0.6.1
basemap==1.0.7
bashplotlib==0.4.4
beanstalkc==0.3.0
beautifulsoup4==4.2.1
biopython==1.61
bleach==1.2.2
blinker==1.2
boto==2.3.0
bottle==0.11.6
bpython==0.12
bzr==2.6.0
census==0.5
chardet==2.2.1
configobj==4.7.2
coverage==3.6
cssselect==0.8
cvxopt==1.1.6
cypari==1.1
dataset==0.4.0
decorator==3.4.0
distro-info==0.0.0
django-blog-zinnia==0.12.3
django-classy-tags==0.4
django-cms==2.4.2
django-debug-toolbar==1.1
django-disqus==0.4.1
django-ganalytics==0.2
-e git://github.com/thornomad/django-hitcount.git@178680e43766982f5ebf05b6b5620e9fea70b37c#egg=django_hitcount-master
django-mptt==0.5.5
django-openid-auth==0.5
django-registration==0.8
django-sekizai==0.7
django-social-auth==0.7.23
django-tagging==0.3.1
django-verbatim==0.1
django-xmlrpc==0.1.5
django4facebook==0.1.0
djangorestframework==2.3.13
docopt==0.6.1
docutils==0.10
dropbox==1.5.1
dulwich==0.9.4
epydoc==3.0.1
facebook-sdk==0.4.0
feedgenerator==1.6
feedparser==5.1.3
filebrowser-safe==0.2.27
futures==2.1.3
gdata==2.0.17
geopy==0.95.1
gevent==0.13.8
geventhttpclient==1.0a
gitdb==0.5.4
gmpy==1.16
google-api-python-client==1.1
goose-extractor==1.0.8
grappelli-safe==0.2.21
greenlet==0.4.1
grequests==0.2.0
grok==1.11.3
grokcore.annotation==1.3
grokcore.catalog==2.0
grokcore.chameleon==1.0.3
grokcore.component==2.5
grokcore.content==1.1
grokcore.formlib==1.9
grokcore.json==1.2
grokcore.layout==1.6
grokcore.message==0.4.2
grokcore.rest==1.3
grokcore.security==1.6.2
grokcore.site==1.6.1
grokcore.traverser==1.2
grokcore.view==2.8
grokcore.viewlet==1.11
grokcore.xmlrpc==1.2
hg-git==0.4.0
html5lib==0.95
http-parser==0.8.1
httplib2==0.8
incf.countryutils==1.0
iotop==0.6
ipython==1.0.0
isodate==0.4.9
jellyfish==0.2.0
jieba==0.32
la==0.6.0
lxml==3.3.4
marisa-trie==0.5.1
martian==0.14
matplotlib==1.3.1
mechanize==0.2.5
mercurial==2.6.3
mlpy==3.5.0
mock==1.0.1
mod-pywebsocket==0.6b1
moin==1.9.7
mplh5canvas==0.7
mpmath==0.17
mysql-connector-python==1.0.12
netifaces==0.8
networkx==1.7
nltk==2.0.4
nose==1.3.0
numexpr==2.4
numpy==1.8.1
oauth==1.0.1
oauth2==1.5.211
oauthlib==0.6.1
openopt==0.5304
openpyxl==1.6.2
oursql==0.9.3.1
pandas==0.13.1
paramiko==1.10.1
patsy==0.2.1
pdflatex==0.1.0
pdfminer==20110515
peewee==2.1.2
pelican==3.2.2
pep8==1.4.5
persistent==4.0.6
pexpect==2.4
pisa==3.0.33
plink==1.6
pp==1.6.4
prettytable==0.6.1
psutil==1.2.1
psycopg2==2.5
pudb==2013.3.2
pyDatalog==0.13.0
pyOpenSSL==0.13
pyUnRAR2==0.99.3
pycrypto==2.6
pycsp==0.9.0
pycurl==7.19.0
pydot==1.0.28
pyenchant==1.6.5
pyfits==3.1.2
pyflakes==0.7.2
pygal==1.2.1
pygraphviz==1.1
pyhdf==0.8.3
pymc==2.3.2
pymongo==2.5.2
pyparsing==2.0.1
pypdfocr==0.7.4
pypng==0.0.16
pyproj==1.9.3
pyquery==1.2.4
pyramid==1.4.2
pyserial==2.6
python-Levenshtein==0.10.2
python-amazon-product-api==0.2.8
python-dateutil==2.2
python-gflags==2.0
python-inotify==0.5
python-instagram==0.8.0
python-ldap==2.4.13
python-linkedin==4.0
python-magic==0.4.3
python-mcrypt==1.1
python-mhash==1.4
python-openid==2.2.5
python-slugify==0.0.7
pyttk==0.3.2
pytz==2014.2
pyzmq==13.1.0
qrcode==2.7
rdflib==4.0.1
redis==2.8.0
reportlab==2.7
repoze.lru==0.6
requests==2.2.0
requests-cache==0.4.4
requests-oauthlib==0.4.0
resolver-test==1.0.1
restkit==4.2.1
rpy2==2.3.6
s3cmd==1.1.0-beta3
scikit-learn==0.14.1
scikits-image==0.7.1
scikits.statsmodels==0.3.1
scipy==0.13.3
selenium==2.33.0
setproctitle==1.1.7
simplejson==3.3.0
six==1.2.0
smmap==0.8.2
snappy==2.1
soaplib==2.0.0-beta2
socketpool==0.5.2
sockjs-tornado==1.0.0
speaklater==1.3
spectrum==0.5.6
spherogram==1.3
splinter==0.5.2
sqlalchemy-migrate==0.8.2
sqlparse==0.1.11
ssh-import-id==3.14
statsmodels==0.5.0
stevedore==0.9.1
stripe==1.12.0
suds==0.4
sympy==0.7.2
tables==3.1.1
texcaller==0
tornado==3.0.2
transaction==1.4.1
translationstring==1.1
tweepy==2.0
twilio==3.5.1
twitter==1.9.4
unittest2==0.5.1
urllib3==1.5
urwid==1.1.1
us==0.7
venusian==1.0a8
virtualenv==1.9.1
virtualenv-clone==0.2.4
virtualenvwrapper==4.0
w3lib==1.3
web.py==0.37
webapp2==2.5.2
wsgiref==0.1.2
xlrd==0.9.2
xlutils==1.6.0
xlwt==0.7.5
xmltodict==0.8.6
z3c.autoinclude==0.3.4
z3c.flashmessage==1.3
z3c.pt==3.0.0a1
zc.buildout==2.1.0
zc.catalog==1.5.1
zc.lockfile==1.1.0
zdaemon==4.0.0
zope.annotation==4.2.0
zope.app.appsetup==4.0.0a1
zope.app.publication==4.0.0a1.dev
zope.app.wsgi==4.0.0a4
zope.authentication==4.1.0
zope.browser==2.0.2
zope.browserpage==4.1.0a1
zope.browserresource==4.0.1
zope.catalog==4.0.0a1
zope.component==4.1.0
zope.configuration==4.0.2
zope.container==4.0.0a3
zope.contentprovider==4.0.0a1
zope.contenttype==4.0.1
zope.datetime==4.0.0
zope.deprecation==4.0.2
zope.dottedname==4.0.1
zope.error==4.1.0
zope.errorview==0.11
zope.event==4.0.2
zope.exceptions==4.0.6
zope.filerepresentation==4.0.2
zope.formlib==4.3.0a1
zope.i18n==4.0.0a4
zope.i18nmessageid==4.0.2
zope.index==4.0.1
zope.interface==4.0.1
zope.intid==4.0.0a1
zope.keyreference==4.0.0a2
zope.lifecycleevent==4.0.2
zope.location==4.0.2
zope.login==2.0.0a1
zope.minmax==2.0.0
zope.pagetemplate==4.0.4
zope.password==4.0.2
zope.principalregistry==4.0.0a2
zope.processlifetime==2.0.0
zope.proxy==4.1.3
zope.ptresource==4.0.0a1
zope.publisher==4.0.0a4
zope.schema==4.3.2
zope.security==4.0.0b1
zope.securitypolicy==4.0.0a1
zope.session==4.0.0a1
zope.site==4.0.0a1
zope.size==4.0.1
zope.tal==4.0.0a1
zope.tales==4.0.1
zope.testing==4.1.2
zope.traversing==4.0.0a3
zope.viewlet==4.0.0a1
